+++ src/runner/index.ts
@@ -1,10 +1,29 @@
 import { Config } from "@jest/types";
 import { readConfig } from "jest-config";
-import { ReporterType, JestReporterType } from "../reporter";
 import { JestRunner } from "./JestRunner";
 import Parser = require("tap-parser");
 import TapMochaReporter = require("tap-mocha-reporter");
 import { TestRunner } from "./TestRunner";
+
+export type MochaReporterType =
+    | "classic"
+    | "doc"
+    | "dot"
+    | "dump"
+    | "json"
+    | "jsonstream"
+    | "landing"
+    | "list"
+    | "markdown"
+    | "min"
+    | "nyan"
+    | "progress"
+    | "silent"
+    | "spec"
+    | "tap"
+    | "xunit";
+export type JestReporterType = "jest" | "jest-verbose";
+export type ReporterType = JestReporterType | MochaReporterType;
 
 function isJestReporter(
     reporterName: ReporterType
